rework to use Workload Identity (#5)
Heavily revise the repo to use Nomad's Workload Identity. From a high-level
perspective, we use WI to authorize the gateway job to gateway job to correctly
register itself with Consul. The API Gateway is deployed in its own Nomad
namespace. You'll add a Consul ACL role that the Consul binding rule matches for
that Nomad namespace. And then that Consul ACL role grants the appropriate
permissions to the API Gateway.

This also refactors the repo to split out the setup of the agents from the API
GW setup, because a user may have a running cluster already either for
production purposes or one they stood up using the tutorial for Nomad and Consul
Workload Identity.

Fixes: https://github.com/hashicorp/consul-gateway-on-nomad/issues/2
Fixes: https://github.com/hashicorp/consul-gateway-on-nomad/issues/3